+++ Task1/Task1/Task1.cs
@@ -1,143 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Text.RegularExpressions;
 
-namespace AS_Course
+namespace Task1
 {
-    abstract class TextProcessor
-    {
-        public static string getText(string path)
-        {
-            if (!File.Exists(path))
-            {
-                throw new Exception("This file does not exist");
-            }
-            return File.ReadAllText(path);
-        }
-    }
-
-    abstract class TextDeleter : TextProcessor
-    {
-        public static void DeleteSubstring(string originalPath)
-        {
-            string originalText = getText(originalPath);
-
-            Console.WriteLine("What to delete: ");
-            string deletingString = Console.ReadLine();
-
-            if (originalText.IndexOf(deletingString) < 0) Console.WriteLine("There is no such text!");
-
-            string copyPath = Path.ChangeExtension(originalPath, ".bak" + Path.GetExtension(originalPath));
-            File.WriteAllText(copyPath, originalText);
-
-            originalText = originalText.Replace(deletingString, "");
-            File.WriteAllText(originalPath, originalText);
-            Console.WriteLine("Done!");
-        }
-    }
-
-    abstract class WordCounter : TextProcessor
-    {
-        public static void Count(string path)
-        {
-            Match words = Regex.Match(getText(path), @"\b\w+[-']*\w*\b");
-            int counter = 0;
-            string selectedWords = "";
-            while (words.Success)
-            {
-                counter++;
-                if (counter % 10 == 0)
-                {
-                    selectedWords += words.Value + ' ';
-                }
-                words = words.NextMatch();
-            }
-
-            Console.WriteLine("There is {0} words in this text, here is every 10th word:\n{1}", counter, selectedWords);
-        }
-    }
-
-    abstract class SentanceReverser : TextProcessor
-    {
-        public static void Reverse(string path, int sentanceNumber)
-        {
-            Match sentances = Regex.Match(getText(path), @"[A-Z].*?[\.!?]");
-            while (sentances.Success)
-            {
-                sentanceNumber--;
-                if (sentanceNumber == 0)
-                {
-                    Match words = Regex.Match(sentances.Value, @"\b\w+[-']*\w*\b");
-                    Match notWords = Regex.Match(sentances.Value, @"\W+");
-                    string Result = "";
-                    while (words.Success)
-                    {
-                        char[] word = words.Value.ToCharArray();
-                        Array.Reverse(word);
-                        if (Result.Length == 0) Result = new string(word);
-                        else
-                        {
-                            Result += notWords.Value + new string(word);
-                            notWords = notWords.NextMatch();
-                        }
-                            words = words.NextMatch();
-                    }
-                    Result += notWords.Value;
-                    Console.WriteLine(Result);
-                    return;
-                }
-                sentances = sentances.NextMatch();
-            }
-        }
-    }
-
-    abstract class DirectoryNavigator
-    {
-        class Comp : IComparer<FileSystemInfo>
-        {
-            public int Compare(FileSystemInfo x, FileSystemInfo y)
-            {
-                return String.Compare(x.Name, y.Name);
-            }
-        }
-        public static void StartNavigation(string path)
-        {
-            Console.Clear();
-            Console.WriteLine("Welcome to Directory Navigation!");
-            Console.WriteLine(path);
-            DirectoryInfo curDir = new DirectoryInfo(path);
-            var filesAndFolders = curDir.GetFileSystemInfos();
-            Console.WriteLine("\nID | Name");
-            int i = 1;
-            Array.Sort(filesAndFolders, new Comp());
-            foreach (var p in filesAndFolders)
-            {
-                Console.WriteLine(i + " | " + p.Name);
-                i++;
-            }
-            Console.WriteLine("Enter ID (0 to go up and everything else to exit):");
-            try
-            {
-                i = Convert.ToInt32(Console.ReadLine());
-            }
-            catch
-            {
-                Console.WriteLine("Not an ID. Exiting");
-            }
-            if (i == 0) StartNavigation(curDir.Parent.FullName);
-            foreach (var p in filesAndFolders)
-            {
-                i--;
-                if (i == 0)
-                {
-                    if (p.Name.IndexOf('.') < 0) StartNavigation(p.FullName);
-                    else Console.WriteLine("It's a file!");
-                }
-            }
-        }
-    }
-
     class Task1
     {
         static void Main(string[] args)
